depthpro_distillation/requirements.txt
matplotlib==3.10.1
numpy==2.2.5
Pillow==11.2.1
pillow_heif==0.22.0
timm==1.0.15
torch==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
